Add files via upload
This HTML file is a demo with instructions on how to classify five drone images of Heron Reef in Google Earth Engine. It includes instructions for accuracy assessment as well as how to measure classified areas per class per drone image (and export as a .CSV file) and how to export single classified images.